pyproject.toml
[tool.poetry]
name = "dj-prego"
version = "0.1.0"
description = "Django authentication pregoing, with REACT!!"
authors = ["syafiq <ahmadsyafiq93@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.5"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.9.1"
celery = "^5.1.2"
redis = "^3.5.3"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
ipython = "^7.25.0"
pytest = "^6.2.4"
pytest-django = "^4.4.0"
factory-boy = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
